extract index functions into files
