+++ src/cli/abort.ts
@@ -1,0 +1,6 @@
+import chalk from "chalk";
+
+export const abortCLI = () => {
+  console.log(chalk.red("Aborting ..."));
+  process.exit(1);
+};
