+++ src/util/getPaths.ts
@@ -1,0 +1,10 @@
+import path from "path";
+import { root } from "./dirname.js";
+
+export const getPaths = (projectName = "") => {
+  return {
+    cwd: process.cwd(),
+    targetDir: path.join(process.cwd(), projectName),
+    templateDir: path.join(root, "template"),
+  };
+};
